Reorganize project directories and move files for a cleaner structure 🧹🧹
